fix(frontend): ThemeContext - move utils and useCallback toggle to satisfy Fast Refresh and lint
